+++ apps/web/src/app/[locale]/shop/_auction/Background.tsx
