+++ apps/web/src/app/[locale]/shop/_auction/DefaultTabRight.tsx
