+++ apps/web/src/app/[locale]/shop/_auction/PersonaSearch.tsx
