+++ apps/web/src/app/[locale]/shop/_auction/ProductTable.tsx
