+++ apps/web/src/app/[locale]/shop/_auction/SellSection/PetList.tsx
