+++ apps/web/src/app/[locale]/shop/_auction/table.styles.ts
