+++ apps/web/src/app/[locale]/shop/_background/BackgroundSection.tsx
