+++ apps/web/src/app/[locale]/shop/_common/FloatingPointSection.tsx
