+++ apps/web/src/app/[locale]/shop/_petGotcha/TenPet.tsx
