+++ apps/web/src/app/[locale]/shop/_petGotcha/useCheckEnoughMoney.ts
