+++ apps/web/src/app/[locale]/shop/page.tsx
@@ -1,14 +1,11 @@
 import { notFound } from 'next/navigation';
-import { getTranslations } from 'next-intl/server';
-import { css } from '_panda/css';
-import { center } from '_panda/patterns';
 
 import GNB from '@/components/GNB/GNB';
 
-import { BackgroundSection } from './BackgroundSection/BackgroundSection';
-import { FloatingPointSection } from './FloatingPointSection/FloatingPointSection';
-import { AuctionSection } from './AuctionSection';
-import { PetGotcha } from './PetGotcha';
+import { BackgroundSection } from './_background/BackgroundSection';
+import { FloatingPointSection } from './_common/FloatingPointSection';
+import { AuctionSection } from './_auction';
+import { PetGotcha } from './_petGotcha';
 import type { TabType } from './type';
 import { TABS } from './type';
 
@@ -20,7 +17,6 @@ async function ShopPage({
   };
 }) {
   const searchParamsTab = searchParams.tab ?? 'products';
-  const t = await getTranslations('Shop');
 
   // NOTE: 탭이 없을 때는 기본값으로 products를 사용
   if (!TABS.includes(searchParamsTab)) {
@@ -28,40 +24,17 @@ async function ShopPage({
   }
 
   return (
-    <>
-      <div className={subStyle}>
-        <GNB />
-        <FloatingPointSection />
+    <div>
+      <GNB />
+      <FloatingPointSection />
 
-        <main>
-          <PetGotcha />
-          <AuctionSection selectedTab={searchParamsTab} />
-          <BackgroundSection />
-        </main>
-      </div>
-      {/* <div className={noticeStyle}>{t('no-mobile-support')}</div> */}
-    </>
+      <main>
+        <PetGotcha />
+        <AuctionSection selectedTab={searchParamsTab} />
+        <BackgroundSection />
+      </main>
+    </div>
   );
 }
 
 export default ShopPage;
-
-const subStyle = css({});
-
-const noticeStyle = center({
-  position: 'fixed',
-  top: 0,
-  left: 0,
-  width: '100%',
-  height: '100%',
-  background: '#454545',
-  color: 'white',
-  zIndex: 'modal',
-  display: 'none',
-  textStyle: 'glyph24.bold',
-  lineHeight: '1.5',
-  textAlign: 'center',
-  _mobile: {
-    display: 'flex',
-  },
-});
